mcq styling
