+++ cassandratin13/mcq_plugin/mcq.css
@@ -3,17 +3,16 @@
     --light-grey: #eeeeee;
     --light-blue: #c1dafd;
     --med-blue: #2858b3;
-    --dark-blue: #123767;
     --darker-blue: #0c2654;
     --green: #9ae9b3;
     --dark-green: #0a6000;
     --light-red: #ff9494;
     --dark-red: #8f0000;
     --font-size: 1.1rem;
+    padding: 1.5rem;
 }
 
 .mc .question {
-    font-size: var(--font-size);
     padding: 2px;
     margin-bottom: 1rem;
 }
@@ -32,9 +31,9 @@
     cursor: pointer;
 
     &:hover {
-        background-color: var(--light-grey);
-        transition: background-color 0.2s ease;
-    }
+        border: 2px solid var(--dark-blue);
+        transition: border 0.1s ease;
+      }
 
     &:has(input[type="radio"]:checked) {
         background-color: var(--light-blue);
@@ -51,7 +50,7 @@
 
 .mc input[type="radio"] {
     margin-top: 1px;
-    margin-right: 0.75rem;
+    margin-right: 1rem;
     margin-left: 0.25rem;
     vertical-align: middle;
     position: relative;
@@ -72,7 +71,7 @@
 .mc input[type="radio"]:before {
     width: 1rem;
     height: 1rem;
-    border-width: 2px;
+    border-width: 1px;
     border-color: #7a7a7a;
     background-color: white;
     z-index: 1;
@@ -119,25 +118,6 @@
 
 .mc .disabled > input[type="radio"] {
     visibility: hidden;
-}
-
-/* Check button and results */
-.mc .mc-button {
-    background-color: var(--med-blue);
-    color: white;
-    border-radius: 18px;
-    display: inline-flex;
-    padding: 0.4rem 1.6rem 0.4rem 1.5rem;
-    margin: 10px 0;
-
-    &:hover {
-        background-color: var(--dark-blue);
-        transition: background-color 0.2s ease;
-    }
-
-    &:disabled {
-        display: none;
-    }
 }
 
 .mc .checkmark-circle {
