+++ cassandratin13/sort_paragraphs_plugin/stp.css
@@ -4,7 +4,6 @@
   --light-blue: #c1dafd;
   --lightest-blue: #d3e5ff;
   --med-blue: #2858b3;
-  /* --dark-blue: #123767; */
   --darker-blue: #0c2654;
   --green: #9ae9b3;
   --dark-green: #0a6000;
